new Mentor Inventor examples transformed to python -> added necessary files with appropriate glue code.
added workaround with %typemap(out) for the %apply directive which got
not applied on getValue(fff[f]) for SbVec3f.h and SbVec4f.h. (don't have
a clue why it works for SbVec2f.h, though. bug(tm) in swig?)

listening to "Not an Addict"! "Breathing iiiiin and breathing ooooooout..."