the first brick for the pyqt bridge is deployed. from the comments:
the next stunt here deserves documentation as otherwise i would not
know what is going on here by tomorrow morning!

What are we doing here? hacking in extremo! i had to find a way to pass
the QEvent instance we get to PyQt!

the approach i chose is to create a QEvent instance in Python from PyQt
(obviously to let this work the user had to import PyQt beforehand. he
has to otherwise he has no reason to create this callback).
this gives us a fully and properly instantiated structure as PyQt
expects it without digging into sip or depending on the sip library.

then i pass the instantiated structure over here and grab the sipThis
entry from the instance which holds a u.cppPtr which turns
out to be the real Qt Object in memory. *har, we are in business now*

now i delete the current pointer and let u.cppPtr point to our very own
QEvent Object. amazingly enough this really works...