build system "fixes" for windows...
strategy:
---------
we need the runtime swig library built as a dll. unfortunately the
precompiled SWIG windows distribution does not come with one. so in order
to avoid in having to figure out how to build and deploy on the crap
called windows myself (and in order to save from a myriad of other
braindead windows issues which will cost me even more time and health)
i simply let distutils believe the swig runtime sources are just another
python extension (i therefore hacked in void init_swigpy() {} into
pivy_swigpy.c). additionally in order to not clash with a probably already
installed swigpy.pyd i renamed it to pivy_swigpy.

i define a PIVY_WIN32 macro in order to fix certain issues such as not
provided DLL export macros in the Coin headers and other windows specific
braindeadness.

another problem:
----------------
regarding the "Python was built with version 6 of Visual Studio and
extensions need to be built with the same version of the compiler,
but it isn't installed." error that distutils correctly issues, you have
2 options to choose from:
1. easy going: uncomment the check in distutils/msvccompiler.py. pivy works
   but then avoid or test thoroughly any usage of methods or functions that
   involve 'FILE *' references. for whatever reason but microstupid figured
   that they need to change the FILE structure between the 2 releases.
2. the hard but safe way: recompile python with VC7 and use this freshly
   generated interpreter for pivy. i am not aware of a precompiled python
   distribution that was created with VC7.

in any case you should play safe and recompile Coin with VC7 which works
out of the box without any glitch just by using the provided project files.