Simplify and generalize coin library detection for win32.
Do not automatically probe for Coin debug libraries as the
majority of Python installations are linked against the release
runtime and will cause hard to track down crashes upon importing
Pivy linked against a Coin debug library. Better to let the user
patch this manually or provide an env var override in the future
for the uncommon case.