fix packages search on Unix platforms
The CMakeLists.txt uses find_package() macro in CONFIG mode to detect coin and soqt packages information. These files on Unix platforms are normally searched in 

<prefix>/(lib/<arch>|lib*|share)/cmake/<name>*/ 
<prefix>/(lib/<arch>|lib*|share)/<name>*/ 
<prefix>/(lib/<arch>|lib*|share)/<name>*/(cmake|CMake)/

If the project type is NONE those paths are excluded from the search. Let's use the default project type which is C and CXX.